Check for configs before writing new ones
Do not overwrite configurations if they already exist

+++ install/00_homedir-setup.sh
@@ -6,7 +6,11 @@ mkdir -p ~/bin
 # Basic Git Config
 git_config="./config/gitconfig"
 if [ -f $git_config ]; then
-	cp $git_config ~/.gitconfig
+	if [ ! -f $HOME/.gitconfig ]; then
+		cp $git_config ~/.gitconfig
+	else
+		printf "\n\nExisting git configuration found\n\n"
+	fi
 else
 	printf "\n\n\Git config not found.\n\n"
 fi
